pyproject.toml
[project]
name = "teams-transcript-notion-sync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "kichinosukey", email = "kichinosuke.fukuhara@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "notion-client>=2.7.0",
    "openai>=2.9.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[project.scripts]
teams-transcript-notion-sync = "teams_transcript_notion_sync:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.workspace]
members = ["teams-transcript-notion-sync"]
